# api/swagger/swagger.yaml
basePath: /api/1.0/
definitions:
...
  github_com_openinfradev_tks-api_pkg_domain.AddOrganizationStackTemplatesRequest:
    properties:
      stackTemplateIds:
        items:
          type: string
        type: array
    required:
    - stackTemplateIds
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.CheckStackTemplateNameResponse:
    properties:
      existed:
        type: boolean
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.CreateStackTemplateRequest:
    properties:
      cloudService:
        enum:
        - AWS
        - BYOH
        type: string
      description:
        type: string
      kubeType:
        type: string
      kubeVersion:
        type: string
      name:
        type: string
      organizationIds:
        items:
          type: string
        type: array
      platform:
        type: string
      serviceIds:
        items:
          type: string
        type: array
      template:
        type: string
      templateType:
        enum:
        - STANDARD
        - MSA
        type: string
      version:
        type: string
    required:
    - kubeType
    - kubeVersion
    - name
    - organizationIds
    - platform
    - serviceIds
    - template
    - version
    type: object
  github_com_openinfradev_tks-api_pkg_domain.CreateStackTemplateResponse:
    properties:
      id:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.FilterResponse:
    properties:
      column:
        type: string
      operator:
        type: string
      releation:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.GetStackTemplateResponse:
    properties:
      stackTemplate:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.StackTemplateResponse'
    type: object
  github_com_openinfradev_tks-api_pkg_domain.GetStackTemplateServicesResponse:
    properties:
      services:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.StackTemplateServiceResponse'
        type: array
    type: object
  github_com_openinfradev_tks-api_pkg_domain.GetStackTemplateTemplateIdsResponse:
    properties:
      templateIds:
        items:
          type: string
        type: array
    type: object
  github_com_openinfradev_tks-api_pkg_domain.GetStackTemplatesResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.PaginationResponse'
      stackTemplates:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.StackTemplateResponse'
        type: array
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.PaginationResponse:
    properties:
      filters:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.FilterResponse'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sortColumn:
        type: string
      sortOrder:
        type: string
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.RemoveOrganizationStackTemplatesRequest:
    properties:
      stackTemplateIds:
        items:
          type: string
        type: array
    required:
    - stackTemplateIds
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.SimpleOrganizationResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.SimpleUserResponse:
    properties:
      accountId:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.StackTemplateResponse:
    properties:
      cloudService:
        type: string
      createdAt:
        type: string
      creator:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleUserResponse'
      description:
        type: string
      id:
        type: string
      kubeType:
        type: string
      kubeVersion:
        type: string
      name:
        type: string
      organizations:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleOrganizationResponse'
        type: array
      platform:
        type: string
      services:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.StackTemplateServiceResponse'
        type: array
      template:
        type: string
      templateType:
        type: string
      updatedAt:
        type: string
      updator:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleUserResponse'
      version:
        type: string
    type: object
  github_com_openinfradev_tks-api_pkg_domain.StackTemplateServiceApplicationResponse:
    properties:
      description:
        type: string
      name:
        type: string
      version:
        type: string
    type: object
  github_com_openinfradev_tks-api_pkg_domain.StackTemplateServiceResponse:
    properties:
      applications:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.StackTemplateServiceApplicationResponse'
        type: array
      type:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.UpdateStackTemplateOrganizationsRequest:
    properties:
      organizationIds:
        items:
          type: string
        type: array
    required:
    - organizationIds
    type: object
  github_com_openinfradev_tks-api_pkg_domain.UpdateStackTemplateRequest:
    properties:
      cloudService:
        enum:
        - AWS
        - BYOH
        type: string
      description:
        type: string
      kubeType:
        type: string
      kubeVersion:
        type: string
      name:
        type: string
      organizationIds:
        items:
          type: string
        type: array
      platform:
        type: string
      serviceIds:
        items:
          type: string
        type: array
      template:
        type: string
      templateType:
        enum:
        - STANDARD
        - MSA
        type: string
      version:
        type: string
    required:
    - kubeType
    - kubeVersion
    - name
    - organizationIds
    - platform
    - serviceIds
    - template
    - version
    type: object
...
host: tks-api-dev.taco-cat.xyz
info:
  contact:
    email: taekyu.kang@sk.com
    name: taekyu.kang@sk.com
  description: This is backend api service for tks platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: tks-api service
  version: "1.0"
paths:
...
  /admin/stack-templates:
    get:
      consumes:
      - application/json
      description: Get StackTemplates
      parameters:
      - description: pageSize
        in: query
        name: pageSize
        type: string
      - description: pageNumber
        in: query
        name: pageNumber
        type: string
      - description: sortColumn
        in: query
        name: soertColumn
        type: string
      - description: sortOrder
        in: query
        name: sortOrder
        type: string
      - collectionFormat: csv
        description: filters
        in: query
        items:
          type: string
        name: filters
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetStackTemplatesResponse'
      security:
      - JWT: []
      summary: Get StackTemplates
      tags:
      - StackTemplates
    post:
      consumes:
      - application/json
      description: Create StackTemplate
      parameters:
      - description: create stack template request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.CreateStackTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.CreateStackTemplateResponse'
      security:
      - JWT: []
      summary: Create StackTemplate
      tags:
      - StackTemplates
  /admin/stack-templates/{stackTemplateId}:
    delete:
      consumes:
      - application/json
      description: Delete StackTemplate
      parameters:
      - description: stackTemplateId
        in: path
        name: stackTemplateId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Delete StackTemplate
      tags:
      - StackTemplates
    get:
      consumes:
      - application/json
      description: Get StackTemplate
      parameters:
      - description: stackTemplateId
        in: path
        name: stackTemplateId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetStackTemplateResponse'
      security:
      - JWT: []
      summary: Get StackTemplate
      tags:
      - StackTemplates
    put:
      consumes:
      - application/json
      description: Update StackTemplate
      parameters:
      - description: Update stack template request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.UpdateStackTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Update StackTemplate
      tags:
      - StackTemplates
  /admin/stack-templates/{stackTemplateId}/organizations:
    put:
      consumes:
      - application/json
      description: Update StackTemplate organizations
      parameters:
      - description: Update stack template organizations request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.UpdateStackTemplateOrganizationsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Update StackTemplate organizations
      tags:
      - StackTemplates
  /admin/stack-templates/name/{name}/existence:
    get:
      consumes:
      - application/json
      description: Check name for stackTemplate
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.CheckStackTemplateNameResponse'
      security:
      - JWT: []
      summary: Check name for stackTemplate
      tags:
      - StackTemplates
  /admin/stack-templates/services:
    get:
      consumes:
      - application/json
      description: Get GetStackTemplateServices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetStackTemplateServicesResponse'
      security:
      - JWT: []
      summary: Get GetStackTemplateServices
      tags:
      - StackTemplates
  /admin/stack-templates/template-ids:
    get:
      consumes:
      - application/json
      description: Get GetStackTemplateTemplateIds
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetStackTemplateTemplateIdsResponse'
      security:
      - JWT: []
      summary: Get GetStackTemplateTemplateIds
      tags:
      - StackTemplates
...
  /organizations/{organizationId}/stack-templates:
    get:
      consumes:
      - application/json
      description: Get Organization StackTemplates
      parameters:
      - description: pageSize
        in: query
        name: pageSize
        type: string
      - description: pageNumber
        in: query
        name: pageNumber
        type: string
      - description: sortColumn
        in: query
        name: soertColumn
        type: string
      - description: sortOrder
        in: query
        name: sortOrder
        type: string
      - collectionFormat: csv
        description: filters
        in: query
        items:
          type: string
        name: filters
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetStackTemplatesResponse'
      security:
      - JWT: []
      summary: Get Organization StackTemplates
      tags:
      - StackTemplates
    post:
      consumes:
      - application/json
      description: Add organization stackTemplates
      parameters:
      - description: Add organization stack templates request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.AddOrganizationStackTemplatesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Add organization stackTemplates
      tags:
      - StackTemplates
    put:
      consumes:
      - application/json
      description: Remove organization stackTemplates
      parameters:
      - description: Remove organization stack templates request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.RemoveOrganizationStackTemplatesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Remove organization stackTemplates
      tags:
      - StackTemplates
  /organizations/{organizationId}/stack-templates/{stackTemplateId}:
    get:
      consumes:
      - application/json
      description: Get Organization StackTemplate
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetStackTemplateResponse'
      security:
      - JWT: []
      summary: Get Organization StackTemplate
      tags:
      - StackTemplates
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"